Moved Components from pror CTD-Component folder to Component Folder
