+++ src/components/SideBarCloseButton/index.js
